A streamlit web application for detecting spectacles in live video stream is built. For this project, python, opencv, mtcnn, tensorflow, keras and streamlit library in python are used.
A streamlit web application for detecting spectacles in live video stream is built. For this project, python, opencv, mtcnn, tensorflow, keras and streamlit library in python are used. First, data (basically the images of various people with specs and also images of people without specs, these are basically passport photos or full length images of persons) are collected from internet and other sources. Then, using state-of-the art mtcnn face detection model (Multi-Task Cascaded Convolutional Networks), face is detected, cropped and I saved face images from those people's images and finally used this dataset for the CNN model training for detecting spectacles. Dataset is split into training dataset and test dataset. 80% of that went to training and 20% went to test data set. Total number of images is 500. I built a CNN model and trained it for 30 epochs. Accuracy in the training set was 99.25% and in the test set accuracy was 99%. After training, model was saved and then I built one simple wep app using Opencv and streamlit library which will detect spectacles that a person is wearing in live video stream.